feat(gpu_prover): new logic for rotation of trace buffers (#96)
## What ❔

This PR implements a new logic for the rotation of trace buffers.

## Why ❔

The old way had separate buffers allocated for each type of circuit, as
more circuit will be introduced, this will not work well, therefore a
new logic is implemented that is using allocators with unified size for
all types of buffer allocations.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted.