feat(gpu_prover): coset and tree recomputations (#120)
## What ❔

This PR implements the re-computation capability for cosets and trees
inside the proving workflow.

## Why ❔

Re-conputing cosets and/or trees for prover stages instead of keeping
them trades memory usage for performance allows us to trade VRAM usage
for performance.
With this capability we are able to prove the reduced risc-v 2^23
circuit on GPUs with 24GB VRAM with a small performance penalty
(estimated to be on the order of 10% on the L4 GPU).
With this feature we can switch our recursion strategy to exclusive use
of the 2^23 sized reduced risc-v circuit and therefore improve the
reduction performance significantly and reduce recursion complexity.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted.