feat: Add capability to merge multiple fris (#141)
Previously, option 4 allows for merging FRIs 1 by 1 (2 at a time). This
change allows merging multiple FRIs at a time. This is very useful as it
reduces overall merging time from O(n) to O(logn). The proving pipeline
is done once for all proofs together, instead of once per proof (by
chaining it to the end of the rest).

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>